+++ Online-Shopping-System/Models/Shipping/ShippingType.cs
@@ -5,7 +5,7 @@
         public int ShippingTypeId { get; set; }
         // ShippingTypes: Standard - Express - SameDay
         public string ShippingName { get; set; } = "Standard";
-        public double ShippingCost { get; set; }
+        public decimal ShippingCost { get; set; }
         public int ShippingDuration { get; set; } // in days
 
         public ShippingRecords ShippingRecords { get; set; }
